refactor: use expression impl
