+++ exact-numbers/src/main/kotlin/xyz/lbres/exactnumbers/expressions/ExpressionImpl.kt
@@ -1,0 +1,4 @@
+package xyz.lbres.exactnumbers.expressions
+
+// internal implementation of expression
+internal abstract class ExpressionImpl : Expression()
