+++ exact-numbers/src/main/kotlin/xyz/lbres/exactnumbers/expressions/term/Term.kt
@@ -1,7 +1,7 @@
 package xyz.lbres.exactnumbers.expressions.term
 
 import xyz.lbres.exactnumbers.exactfraction.ExactFraction
-import xyz.lbres.exactnumbers.expressions.expression.Expression
+import xyz.lbres.exactnumbers.expressions.Expression
 import xyz.lbres.exactnumbers.irrationals.IrrationalNumber
 import xyz.lbres.exactnumbers.irrationals.log.Log
 import xyz.lbres.exactnumbers.irrationals.pi.Pi
